Added a block to allow a box to slip over the pin
In the trunk of our Caddy Max we have a rail with 2 pins on each side
to fix the the rear seats. When we do longer drives with a lot of luggage
I remove the last row of seats. Those pins do avaid easy deployment of boxes
because thez do not slip over those pins.

Signed-off-by: Christian Ege <k4230r6@gmail.com>